Move Symbolic subfolder out of Term namespace
