+++ src/Pirouette.hs
@@ -1,6 +1,6 @@
 module Pirouette (module X) where
 
-import Pirouette.Term.Symbolic.Prover.Runner as X
+import Pirouette.Symbolic.Prover.Runner as X
   ( assertIncorrectnessLogic,
     replIncorrectnessLogic,
     IncorrectnessParams(..)
