+++ tests/unit/Pirouette/Symbolic/EvalSpec.hs
@@ -4,20 +4,20 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE QuasiQuotes #-}
 
-module Pirouette.Term.SymbolicEvalSpec (tests) where
+module Pirouette.Symbolic.EvalSpec (tests) where
 
 import Control.Monad.Reader
 import Language.Pirouette.Example
 import Pirouette.Monad
 import Pirouette.Term.Syntax.Base
-import Pirouette.Term.Symbolic.Eval
+import Pirouette.Symbolic.Eval
 import Pirouette.Transformations ( elimEvenOddMutRec )
 import Test.Tasty
 import Test.Tasty.HUnit
 import qualified Pirouette.SMT.Constraints as SMT
 import qualified PureSMT as SMT
 
-import Pirouette.Term.SymbolicEvalUtils
+import Pirouette.Symbolic.EvalUtils
 
 {-
 -- TODO: Rewrite these tests or drop them; the rewrite should use one interface only
